Moved network definitions inside 'models'
