+++ datasets/WE_MAML/setting.py
@@ -17,7 +17,7 @@ __C_WEMAML.LABEL_FACTOR = 1
 __C_WEMAML.LOG_PARA = 100.
 
 __C_WEMAML.RESUME_MODEL = ''  # model path
-__C_WEMAML.TRAIN_BATCH_SIZE = 12  # imgs
+__C_WEMAML.TRAIN_BATCH_SIZE = 1  # imgs
 __C_WEMAML.TEST_BATCH_SIZE = 1  # imgs
 
 __C_WEMAML.VAL_BATCH_SIZE = 8  #
